# techforce-pro/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: FieldCommand — Premium fire protection field service management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recurring-schedules:
    get:
      operationId: listRecurringSchedules
      tags: [jobs]
      summary: List all recurring job schedules
      responses:
        "200":
          description: Recurring schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringSchedule"
    post:
      operationId: createRecurringSchedule
      tags: [jobs]
      summary: Create a recurring job schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringSchedule"
  /recurring-schedules/{id}:
    get:
      operationId: getRecurringSchedule
      tags: [jobs]
      summary: Get recurring schedule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recurring schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringSchedule"
    put:
      operationId: updateRecurringSchedule
      tags: [jobs]
      summary: Update recurring schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringSchedule"
    delete:
      operationId: deleteRecurringSchedule
      tags: [jobs]
      summary: Delete recurring schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /recurring-schedules/{id}/pause:
    post:
      operationId: pauseRecurringSchedule
      tags: [jobs]
      summary: Pause a recurring schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringSchedule"
  /recurring-schedules/{id}/resume:
    post:
      operationId: resumeRecurringSchedule
      tags: [jobs]
      summary: Resume a paused recurring schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringSchedule"

  # ── INVOICES ─────────────────────────────────────────────────────────────────
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  schemas:
...
    LineItem:
      type: object
      required: [id, service, description, qty, unitPrice, total]
      properties:
        id:
          type: integer
        service:
          type: string
        description:
          type: string
        qty:
          type: number
        unitPrice:
          type: number
        total:
          type: number
    LineItemInput:
      type: object
      required: [service, qty, unitPrice]
      properties:
        service:
          type: string
        description:
          type: string
        qty:
          type: number
        unitPrice:
          type: number
    Invoice:
      type: object
      required: [id, customerId, customerName, status, issued, due, lineItems, subtotal, total, notes]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        techId:
          type: ["integer", "null"]
        techName:
          type: ["string", "null"]
        jobId:
          type: ["integer", "null"]
        estimateId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        issued:
          type: string
        due:
          type: string
        paidDate:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: number
        taxPct:
          type: number
        total:
          type: number
        notes:
          type: string
        reminderSent:
          type: boolean
    InvoiceInput:
      type: object
      required: [customerId, due, lineItems]
      properties:
        customerId:
          type: integer
        techId:
          type: ["integer", "null"]
        jobId:
          type: ["integer", "null"]
        estimateId:
          type: ["integer", "null"]
        due:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItemInput"
        taxPct:
          type: number
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
        techId:
          type: ["integer", "null"]
        due:
          type: string
        notes:
          type: string
        reminderSent:
          type: boolean
...
    RecurringSchedule:
      type: object
      required: [id, customerId, customerName, serviceType, intervalType, startDate, nextOccurrence, status, revenue, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        employeeId:
          type: ["integer", "null"]
        employeeName:
          type: ["string", "null"]
        serviceType:
          type: string
        intervalType:
          type: string
          enum: [6months, 1year, custom]
        customDays:
          type: ["integer", "null"]
        startDate:
          type: string
        nextOccurrence:
          type: string
        status:
          type: string
          enum: [active, paused]
        revenue:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    RecurringScheduleInput:
      type: object
      required: [customerId, serviceType, intervalType, startDate]
      properties:
        customerId:
          type: integer
        employeeId:
          type: ["integer", "null"]
        serviceType:
          type: string
        intervalType:
          type: string
          enum: [6months, 1year, custom]
        customDays:
          type: ["integer", "null"]
        startDate:
          type: string
        revenue:
          type: number
        notes:
          type: ["string", "null"]
    RecurringScheduleUpdate:
      type: object
      properties:
        employeeId:
          type: ["integer", "null"]
        serviceType:
          type: string
        intervalType:
          type: string
        customDays:
          type: ["integer", "null"]
        nextOccurrence:
          type: string
        revenue:
          type: number
        notes:
          type: ["string", "null"]
...